Add turbulent tecscreen model, collect tecscreen-related functions in losito/lib_tecscreen.py
